requirements-prod.txt
Django<1.10,>=1.9
Markdown==2.6.2
Pillow==2.8.1
Pygments==2.0.2
Werkzeug==0.10.4
argparse==1.2.1
av==0.2.2
distribute==0.6.24
django-bootstrap3==5.4.0
django-debug-toolbar<1.9,>=1.8
django-extensions<1.9,>=1.8
django-mptt<0.9,>=0.8
django-pdb<0.7,>=0.6
django-resumable==0.1
django-secure==1.0.1
gunicorn==19.3.0
hamlpy==0.82.2
ipdb==0.8.1
ipython==3.1.0
jsonfield==1.0.3
psycopg2==2.6
python-magic==0.4.12
raven==5.3.1
six==1.9.0
sqlparse==0.1.15
wsgiref==0.1.2
